shieldgemma2/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt --emit-index-url pyproject.toml
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu128

accelerate==1.8.1
    # via shieldgemma2 (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==2.1.1
    # via requests
click==8.2.1
    # via uvicorn
docker==7.1.0
    # via google-cloud-aiplatform
docstring-parser==0.16
    # via google-cloud-aiplatform
dotenv==0.9.9
    # via shieldgemma2 (pyproject.toml)
fastapi==0.114.0
    # via google-cloud-aiplatform
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
google-api-core==2.25.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform==1.100.0
    # via shieldgemma2 (pyproject.toml)
google-cloud-bigquery==3.30.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.14.2
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via google-cloud-aiplatform
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.23.0
    # via google-cloud-aiplatform
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-resource-manager
grpcio==1.73.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.73.1
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   google-cloud-aiplatform
    #   google-genai
huggingface-hub==0.33.1
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==2.1.2
    # via
    #   accelerate
    #   shapely
    #   transformers
nvidia-cublas-cu12==12.8.3.14
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.57
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.61
    # via torch
nvidia-cuda-runtime-cu12==12.8.57
    # via torch
nvidia-cudnn-cu12==9.7.1.26
    # via torch
nvidia-cufft-cu12==11.3.3.41
    # via torch
nvidia-cufile-cu12==1.13.0.11
    # via torch
nvidia-curand-cu12==10.3.9.55
    # via torch
nvidia-cusolver-cu12==11.7.2.55
    # via torch
nvidia-cusparse-cu12==12.5.7.53
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.8.61
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.8.55
    # via torch
packaging==24.1
    # via
    #   accelerate
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   huggingface-hub
    #   transformers
pillow==11.0.0
    # via shieldgemma2 (pyproject.toml)
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psutil==7.0.0
    # via accelerate
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.1.1
    # via
    #   dotenv
    #   uvicorn
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.28.1
    # via
    #   docker
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
    #   huggingface-hub
    #   transformers
rsa==4.9.1
    # via google-auth
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
sentencepiece==0.2.0
    # via shieldgemma2 (pyproject.toml)
setuptools==70.2.0
    # via
    #   torch
    #   triton
shapely==2.1.1
    # via google-cloud-aiplatform
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.38.6
    # via
    #   fastapi
    #   google-cloud-aiplatform
sympy==1.13.3
    # via torch
tenacity==8.5.0
    # via google-genai
tokenizers==0.21.2
    # via transformers
torch==2.7.1+cu128
    # via
    #   shieldgemma2 (pyproject.toml)
    #   accelerate
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   transformers
transformers==4.53.0
    # via shieldgemma2 (pyproject.toml)
triton==3.3.1
    # via torch
typing-extensions==4.12.2
    # via
    #   fastapi
    #   google-cloud-aiplatform
    #   google-genai
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==1.26.13
    # via
    #   docker
    #   requests
uvicorn==0.35.0
    # via google-cloud-aiplatform
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   google-genai
    #   uvicorn
